Update methods in Either class for combining Eithers
Expanded the functionality of the Either class by adding methods that combine or accumulate multiple Either instances into a single Either. The added methods can handle a varying number of Eithers, applying a transform function to Right values if all the inputs are Right, or accumulating Left values into a NonEmptyList of errors otherwise.