Updated documentation in multiple Java classes
Enhanced the documentation of several Java classes. Added comments to explain the purpose, parameters, and return types of functions and methods. This provides clearer understanding of their expected inputs, outputs and behaviors. Clearer communication of these key aspects of functions and methods will facilitate more effective use and maintenance of the code.