Refactor code and add 'recover' method to Try interface
Code across several files has been refactored for readability and clarity. Moreover, a new 'recover' method has been added to the Try interface, which allows the user to provide a handling function for failures. It will only be called when an exception of a specified class is thrown, otherwise, it will pass through. Adjusted tests to include checks for the new functionality. Also, a deprecated GitHub Action has been removed from the workflow.